+++ NEWS.md
@@ -1,16 +1,25 @@
-## CausalGPS (2021-05-12)
+## CausalGPS 0.2.3 (2021-05-12)
 
 ### Added
+### Changed
+### Fixed
 
+*  Fixed documentations
+
+### Removed
+
+
+
+## CausalGPS 0.2.2 (2021-05-12)
+
+### Added
 * estimate_semi_erf
 * estimate_hr
 
 ### Changed
-
 * Package name: GPSmatching --> CausalGPS 
 
 ### Fixed
-
 ### Removed